Move Appointment folder to pages folder
