+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
-import Appointment from './views/Appointment';
+import Appointment from './pages/Appointment';
 
 function App() {
   return (
